- Updated version 3.0.0 of the REST API specification - Updated pom.xml for version 3.0.0-SNAPSHOT - Updated for version 3.0.0 g2.jar function names ("V2" removal) - Removed parsing of "MATCH_SCORE" field and "matchScore" fields as well - Changed parsing of "UTYPE_CODE" to "USAGE_TYPE"

+++ src/test/java/com/senzing/api/services/EntityDataReadServicesTest.java
@@ -2115,7 +2115,7 @@ public class EntityDataReadServicesTest extends AbstractServiceTest {
       long after = System.nanoTime();
 
       // TODO(barry): remove this extra code
-      int flags = this.entityDataServices.getFlags(
+      long flags = this.entityDataServices.getFlags(
           (forceMinimal == null) ? false : forceMinimal,
           (featureMode != null ? featureMode : WITH_DUPLICATES),
           (withFeatureStats != null ? withFeatureStats : false),
